# mobile/openapi.yaml
openapi: 3.0.3
info:
  title: ViVu App API
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Development
  - url: https://api.vivuapp.vn
    description: Production
...
paths:
...
  /api/v1/places:
    get:
      tags: [Places]
      summary: Search places
      parameters:
        - in: query
          name: keyword
          schema:
            type: string
        - in: query
          name: typeId
          schema:
            type: integer
            format: int64
        - in: query
          name: provinceId
          schema:
            type: integer
            format: int64
        - in: query
          name: districtId
          schema:
            type: integer
            format: int64
        - in: query
          name: wardId
          schema:
            type: integer
            format: int64
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/PlaceStatus"
        - in: query
          name: lat
          schema:
            type: number
            format: double
        - in: query
          name: lng
          schema:
            type: number
            format: double
        - in: query
          name: radius
          schema:
            type: number
            format: double
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Page of places
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponsePlacePage"
  /api/v1/places/nearby:
    get:
      tags: [Places]
      summary: Nearby places
      parameters:
        - in: query
          name: lat
          required: true
          schema:
            type: number
            format: double
        - in: query
          name: lng
          required: true
          schema:
            type: number
            format: double
        - in: query
          name: radius
          schema:
            type: number
            format: double
            default: 5
        - in: query
          name: keyword
          schema:
            type: string
        - in: query
          name: typeId
          schema:
            type: integer
            format: int64
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/PlaceStatus"
      responses:
        "200":
          description: Nearby places
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponsePlaceList"
  /api/v1/places/{id}:
    get:
      tags: [Places]
      summary: Place details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Place
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponsePlaceResponse"
  /api/v1/admin/places:
    post:
      tags: [Places]
      security:
        - bearerAuth: []
      summary: Create place (ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponsePlaceResponse"
  /api/v1/admin/places/{id}:
    patch:
      tags: [Places]
      security:
        - bearerAuth: []
      summary: Update place (ADMIN)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponsePlaceResponse"
  /api/v1/admin/places/{id}/status:
    patch:
      tags: [Places]
      security:
        - bearerAuth: []
      summary: Update place status (ADMIN)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceStatusRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponsePlaceResponse"
  /api/v1/place-suggestions:
    post:
      tags: [PlaceSuggestions]
      security:
        - bearerAuth: []
      summary: Submit place suggestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceSuggestionRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponsePlaceSuggestionResponse"
  /api/v1/me/place-suggestions:
    get:
      tags: [PlaceSuggestions]
      security:
        - bearerAuth: []
      summary: My place suggestions
      parameters:
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/PlaceSuggestionStatus"
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Page of suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponsePlaceSuggestionPage"
  /api/v1/admin/place-suggestions:
    get:
      tags: [PlaceSuggestions]
      security:
        - bearerAuth: []
      summary: All suggestions (ADMIN)
      parameters:
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/PlaceSuggestionStatus"
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Page of suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponsePlaceSuggestionPage"
  /api/v1/admin/place-suggestions/{id}/review:
    patch:
      tags: [PlaceSuggestions]
      security:
        - bearerAuth: []
      summary: Review suggestion (ADMIN)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceSuggestionReviewRequest"
      responses:
        "200":
          description: Reviewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponsePlaceSuggestionResponse"
...
  /ai/itineraries/suggest:
    post:
      tags: [AI]
      summary: Suggest AI itinerary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItinerarySuggestRequest"
      responses:
        "200":
          description: Suggested itinerary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuggestedItinerary"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponseBase:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        result:
          nullable: true
...
    ApiResponsePlacePage:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            result:
              $ref: "#/components/schemas/PagePlaceResponse"
    ApiResponsePlaceList:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            result:
              type: array
              items:
                $ref: "#/components/schemas/PlaceResponse"
    ApiResponsePlaceResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            result:
              $ref: "#/components/schemas/PlaceResponse"
    ApiResponsePlaceSuggestionResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            result:
              $ref: "#/components/schemas/PlaceSuggestionResponse"
    ApiResponsePlaceSuggestionPage:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            result:
              $ref: "#/components/schemas/PagePlaceSuggestionResponse"
...
    PlaceTypeResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        name:
          type: string
        description:
          type: string
    ProvinceResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    DistrictResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        provinceId:
          type: integer
          format: int64
    WardResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        districtId:
          type: integer
          format: int64
    PlaceRequest:
      type: object
      required: [name, typeId]
      properties:
        name:
          type: string
        typeId:
          type: integer
          format: int64
        areaId:
          type: integer
          format: int64
        provinceId:
          type: integer
          format: int64
        districtId:
          type: integer
          format: int64
        wardId:
          type: integer
          format: int64
        addressDetail:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        description:
          type: string
        status:
          $ref: "#/components/schemas/PlaceStatus"
        isOfficial:
          type: boolean
        ownerUserId:
          type: integer
          format: int64
    PlaceStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/PlaceStatus"
    PlaceResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        placeType:
          $ref: "#/components/schemas/PlaceTypeResponse"
        province:
          $ref: "#/components/schemas/ProvinceResponse"
        district:
          $ref: "#/components/schemas/DistrictResponse"
        ward:
          $ref: "#/components/schemas/WardResponse"
        addressDetail:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        description:
          type: string
        status:
          $ref: "#/components/schemas/PlaceStatus"
        isOfficial:
          type: boolean
        ownerUserId:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PlaceSuggestionRequest:
      type: object
      required: [placeName]
      properties:
        placeName:
          type: string
        typeId:
          type: integer
          format: int64
        areaId:
          type: integer
          format: int64
        provinceId:
          type: integer
          format: int64
        districtId:
          type: integer
          format: int64
        wardId:
          type: integer
          format: int64
        addressDetail:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        description:
          type: string
    PlaceSuggestionReviewRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/PlaceSuggestionStatus"
        rejectionReason:
          type: string
        createPlace:
          type: boolean
    PlaceSuggestionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        suggestedById:
          type: integer
          format: int64
        suggestedByName:
          type: string
        placeName:
          type: string
        placeType:
          $ref: "#/components/schemas/PlaceTypeResponse"
        province:
          $ref: "#/components/schemas/ProvinceResponse"
        district:
          $ref: "#/components/schemas/DistrictResponse"
        ward:
          $ref: "#/components/schemas/WardResponse"
        addressDetail:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        description:
          type: string
        status:
          $ref: "#/components/schemas/PlaceSuggestionStatus"
        rejectionReason:
          type: string
        reviewedById:
          type: integer
          format: int64
        reviewedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    ItinerarySuggestRequest:
      type: object
      required: [startLatitude, startLongitude]
      properties:
        mood:
          type: string
        startLatitude:
          type: number
          format: double
        startLongitude:
          type: number
          format: double
        radiusKm:
          type: number
          format: double
        maxStops:
          type: integer
        maxDurationHours:
          type: integer
        placeTypes:
          type: array
          items:
            type: string
    SuggestedItinerary:
      type: object
      properties:
        name:
          type: string
        mood:
          type: string
        stops:
          type: array
          items:
            $ref: "#/components/schemas/SuggestedStop"
        totalDistanceKm:
          type: number
          format: double
        totalDurationMinutes:
          type: integer
        routePolyline:
          type: string
        mapPreviewUrl:
          type: string
    SuggestedStop:
      type: object
      properties:
        sequenceOrder:
          type: integer
        googlePlaceId:
          type: string
        placeName:
          type: string
        address:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        rating:
          type: number
          format: double
        suggestedDurationMinutes:
          type: integer
        aiReason:
          type: string
...
    Pageable:
      type: object
      properties:
        pageNumber:
          type: integer
        pageSize:
          type: integer
        sort:
          $ref: "#/components/schemas/Sort"
        offset:
          type: integer
        paged:
          type: boolean
        unpaged:
          type: boolean
    Sort:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean
    PagePlaceResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/PlaceResponse"
        pageable:
          $ref: "#/components/schemas/Pageable"
        totalPages:
          type: integer
        totalElements:
          type: integer
          format: int64
        last:
          type: boolean
        size:
          type: integer
        number:
          type: integer
        sort:
          $ref: "#/components/schemas/Sort"
        numberOfElements:
          type: integer
        first:
          type: boolean
        empty:
          type: boolean
    PagePlaceSuggestionResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/PlaceSuggestionResponse"
        pageable:
          $ref: "#/components/schemas/Pageable"
        totalPages:
          type: integer
        totalElements:
          type: integer
          format: int64
        last:
          type: boolean
        size:
          type: integer
        number:
          type: integer
        sort:
          $ref: "#/components/schemas/Sort"
        numberOfElements:
          type: integer
        first:
          type: boolean
        empty:
          type: boolean
...
    PlaceStatus:
      type: string
      enum: [ACTIVE, TEMP_CLOSED, CLOSED]
    PlaceSuggestionStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED]
...